backend/requirements.txt
# FastAPI 后端依赖清单
# Python 3.10+

# Web框架
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# HTTP客户端
httpx==0.27.2
aiohttp==3.10.10

# 图像处理
Pillow==11.0.0
opencv-python-headless==4.10.0.84

# 数据验证
pydantic==2.9.2
pydantic-settings==2.6.0

# CORS支持
python-dotenv==1.0.1

# 工具库
python-magic==0.4.27
aiofiles==24.1.0

# 日志
loguru==0.7.2

# 定时任务（临时文件清理）
apscheduler==3.10.4

# 开发工具
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
pyjwt>=2.8.0
